Integrate victor_camera into android build
- Compile & commit an ELF armeabi-v7a object file from the pre-processed .c file.

The .c file was preprocessed with gcc, but we’re using clang, so this
is the easiest way to avoid an awkward out-of-band compile step.
Since the preprocessed C code was developed against android sources, it
might not even compile with clang. Finally, we add the -fPIC flag, since
this code will end up in a shared lib in development/debug builds.

- Move code under androidHAL dir, next to the original android NDK
Camera implementation.

- Create a separate static lib (`fugly_camera`) to commemorate this epic
victor milestone.  Also, this keeps it separate from everything else and
easy to delete.