COZMO-14367 Cleanup process before debugging (#5649)
Before we launch a new cozmoengined process for debugging, we should kill
and existing engine process. This change adds a `killall cozmoengined` before
launching the debugee process.

Redirect io from start-lldb-server script to /dev/null so that it does not
spam to the terminal where the program was started.

Exit with error if lldb-server fails to start.