Specify the files uploaded to Smartling for translation (#5771)
Added support for environment variables to specify the subdirectory and pattern to search for files to upload to Smartling. The subdirectory is "unity" for Cozmo and "resources" for OverDrive (default is the entire Git repo). The pattern is "*en-US/*.json" for Cozmo and "*base/*-strings.json" for OverDrive (default is "*.json" files). The environment variables will get set by TeamCity.