COZMO-14603 cmake anim process (#5)
* Fix a case-sensitivity error in cozmoEngine2.gyp

This was fixed in master, but was possibly lost in a merge.

* [audio] add missing engine_clad dep to multiplexer

Audio multiplexer depends on clad file in the engine. The explicit
dependency was missing. The build usually worked because the overall
implicit ordering was correct, but this was causing periodic failures.

* COZMO-14603 CMake build for anim process

This changeset adds support for building a victor_animator binary
on android, and also the webotsCtrlAnim on mac. It includes scripts
to start and stop the process on android. This work required major cleanup
to the cpplite generation process in cmake, which was doing more
work than necessary.

- Generate lightCubeMessage defs & fix robot include paths

- Symlink build output resources into webots controller dirs

  This ensures that webots will have access to all of the appropriate
  resource files.

* Ensure AnkiLogStringTable exist before running configure