COZMO-14527	Fix crash in Cozmo.Needs.Sparks.UI.SparksDetailModal.HandleSparkEnded (#5805)
If you happened to be playing right at 4AM and then tried to get the current session to log some data, there'd be an exception because the CurrentSession usually just used for Daily Goal's was gone. Now that we don't need to restart for Daily goals, anytime current session is invalid, just create a new one.