Mwillett/cozmo 14438 separate nav map from blockworld (#5816)
Main folder is now just navMap, which contains the interface and factory for any navMap type object. Classes and data objects specific to the current MemoryMap implementation are now in a separate memoryMap folder.

It also further separates the quadTree data structure which handles all the logic for managing a basic QT implementation, with proper expansion/splitting/merging logic as well as polygon insert methods.