Explicitly make asset copy depend on src (#24)
This fixes a bug where changes to assets in resources/ did not cause
them to be copied to the build folder.

reported by @andrew-anki

* Define project file based on CMake generator

Use a case statement to set the project file based on the specified
CMake generator. Bash 3 does not support associative arrays, which was
causing the script to always search for `cozmo.xcodeproj`, forcing configure
to always run for Android builds.