COZMO-14446 Needs levels not correct after corrupted robot save (#5890)
* COZMO-14446 Needs levels not correct after corrupted robot save

The actual repro case for this involves corrupting a robot save (by backgrounding
the app and then quickly killing it, before the robot save completes).  Then you
connect to a different robot, then connect back to the first robot...at that point the
robot has a corrupted save and the device save has been overwritten by the
second robot.  So this fix involves writing out an additional device save whenever
backgrounding, with the serial number embedded in the save file name.  Then
after reading from robot and detecting there is no robot save, and that the device
save does not apply to this robot (via serial number match), we look for that
secondary save file and load it if it exists.

- Pull some device-reading code (and stuff that happens immediately after the read)
  into AttemptReadFromDevice, so it can be called from an additional place
- The 'fix' logic in InitAfterReadFromRobot, in the case where there is no robot save
  but there is a device save:  See if the device save was for the newly-connected
  robot, and if not, look for the alternate save that's specific to that robot, and load it
- In app backgrounding, if we have a valid serial number, write the alternate device
  save file (we've just written the primary save file above, in SetPaused)
- Adjust DeviceHasNeedsState, and ReadFromDevice, to take a filename

* COZMO-14446 pr feedback

- Conditional compile the read/write stats so they don't happen in shipping builds