Mwillett/cozmo 6832 approach cliffs backwards (#5610)
Added functionality to import visually identified edges as potential collisions in the path planner.

Additionally, the planner now adds I higher penalty to potential collisions when reversing, requiring a higher level of certainty of navigable space to execute the path. Ideally this will result in fewer occurrences of Cozmo falling backwards of a table it has either detected with a cliff, or with his camera.

New functionality is disabled by default, but can be enabled by setting the console variable kUseNavMapObstacles to true when testing.