Fix bug in loading voiceCommand_config.json (#45)
Problem:
Log shows the following error message
```
webots.test ERROR - Webots log contained an error:
'[webotsCtrlGameEngine2] (t:01) [Error] DataPlatform.readAsJson (tc0000)
: Failed to read
[/Users/arjun/Code/victor/simulator/controllers/webotsCtrlGameEngine2/resources/assets/voiceCommand/exports/voiceCommand_config.json]'
```

Fix:
The file was missing from the folder in which assets are copied into (as
a build step). Fixed by adding a build rule for that file.

Also the wrong path prefix was corrected in the code base that loads
this file.