Disable firmware update processes in Victor (#49)
Firmware update no longer applies for Victor because engine and firmware
need to be updated at the same time. The code in charge of this needs to
be gutted/reworked entirely.

This change disables some of the firmware update related code, and goes
as far as to prevent errors from arising in the log that happen because
there are missing `.safe` file assets as normal.