COZMO-14293 Code Lab: escape project json (#5936)
Change how ProjectJSON is modified and passed to JavaScript from C#. Previously, the code would escape single quote and double quote for ProjectJSON and then call the JavaScript evaluate method with args passed as parameters.

With this PR, we no longer manually escape any chars. Instead, we build a JSON Object on the C# side, then serialize it and pass it to js, where it gets stringified and passed to Scratch. So we are now taking advantage of JSON's ability to correctly serialize and transfer nearly any unicode char. From the JSON spec: "All Unicode characters may be placed within the quotation marks, except for the characters that must be escaped: quotation mark, reverse solidus, and the control characters (U+0000 through U+001F)."