COZMO-14804: Fix crash in - [Cozmo.Needs.Sparks.UI.SparksDetailModal UpdateButtonState ] (#6045)
Wasn't able to repro the crash, and the crash report doesn't include line numbers so I'm guessing that the crash
is coming from the CozmoSings modal, which inherits from the standard details modal but doesn't define a SparksCost or
SparksSpinnerContainer object. If the crash keeps happening, we can try wrapping more things in null checks.