VIC-58 Rolling Shutter (#70)
The idea for matching ImageIMUData with an image is to use steady_clock timestamps which are the same across processes. ImageIMUData now has a timestamp instead of an imageID and a line2Number and images now have stead_clock timestamps. This makes it much easier to match imu data with a specific image/point in time when the image was being captured. To make things slightly simpler I have removed the time offset set by SyncTime so image timestamps no longer start at 0. This should be easier to add back if we want it...

Rolling shutter correction seems to work fairly well still. Checkerboards are correctly unwarped while rotating. The only issue I've noticed is that cube poses while Victor is rotating aren't super stable, they translate side to side. I'm not quite sure what is causing this...