COZMO-15117 Fixing placeholders in CodeLab strings (#6084)
The CodeLabVerticalBlocks.json strings file has a bunch of numbered placeholders for text to be inserted at runtime, eg. "on cube %2 set %1 to %3", so we're updating 'placeholder_format_custom' to account for that and only that. For now, the only text in that file that Smartling will consider to be placeholder is "%" immediately followed by an integer.
The CodeLabStrings.json strings file has a couple strings with two "%" characters each (not representing any placeholder text) and some text in between them, eg. "the height of Cozmo's lift as a % of its maximum (100%) height", so we're updating 'placeholder_format_custom' to account for that by removing the regex that matches that pattern.