COZMO-11529: Code Lab Save project as Remix (#6073)
Includes work for: COZMO-14475, COZMO-11529, COZMO-14993, COZMO-15109, COZMO-15111

Implements "Remix" button on workspace page that opens a dialog to save the project as a remix.

Partially implements a modal dialog to save a modified sample project as a remix when exiting the project. Michelle to finish setting this up soon.

Adds pencil icon next to project name in vertical and horizonatal workspace for user projects and remixes.

Renames remixed projects in horizontal and vertical to start with "Remix of".

Adds character limit to project name input text box.

Refactor code out of export flow to share with remix flow: FindProjectByUUIDAndCreateCodeLabProject.