COZMO-14994 Throttle Code Lab calls (#6162)
Spamming calls to Unity from JS locks up the WebView, because of the way it generates an iframe to send them. This was only an issue with blocks that immediately complete (as blocks that wait for an action etc. force there to be a decent gap between each message).
To throttle the number of messages from instant-complete blocks we now measure the number of messages sent to Unity in a given time period, and if it goes over a certain threshold (30  messages in 100ms) we sleep for long enough to push that to exactly 30 messages in 100ms - i.e. 1 "instant" message now waits up to 100ms. Programs that don't spam messages without sleeping are unaffected - e.g. the clock example was only 9 messages every 200ms.