Wip/paul t/cozmo 14548 audio choppy (#6217)
* COZMO-14548 DAS client ability to pause uploads

- Simple 'pause' flag that prevents attempting to upload DAS files

* COZMO-14548 Use new DAS uploading pause flag

- Pause uploading DAS files when we connect to the robot
- Resume uploading DAS files when we disconnect from the robot

Of course this does not avoid attempting to upload DAS files when the wifi
is still the robot wifi, but at least we're not trying to upload when we know
for sure we are connected to robot

* COZMO-14548 PR feedback

- Rename DASSetIsPaused -> DASPauseUploadingToServer
- Rename SetIsPaused -> SetIsUploadingPaused
- Rename _isPaused -> _isUploadingPaused
- Do nothing in DasAppender::ConsumeALogFile if uploading is paused
- When un-pausing, immediately try to upload

* COZMO-14548 More PR feedback

- Ensure we're actually transitioning to un-paused (from paused) before doing the
  force flush

* COZMO-14548 Choppy audio additional fix attempt

- In robot destructor, reset the _nvStorageComponent pointer as we do the others
- Make NVStorageComponent destructor clear out the _signalHandles vector, so they
  get unsubscribed from the event manager.  Not sure if this is necessary but I had to
  do something similar with NeedsManager when its lifetime matched that of the robot.

The thinking behind these changes to is that possibly, after robot connection (which
queues a bunch of robot read operations), and then there is a firmware update, and
then a robot connection again, that there may be some stray read operations still
coming in.

* COZMO-14548 Choppy audio: Mini: log format

- Fix the formatting for a log message that we're outputting from Unity.

HOWEVER, as it turns out, the variable being shown in this log message is questionable
and seems to only have non-zero data when we run webots simulator.

* COZMO-14548 Audio choppy: Prevent unneeded and redundant writes to robot

For this change we do not write experiment variations to robot when we don't have
to.  Previously this was being done at least once after robot connection, and also
often twice.

- In the "restore assignment" callback handler, instead of queueing up a write to
  robot, just make a copy of the assignments
- In NeedsManager::InitAfterReadFromRobot, after we attempt to activate both
  experiments, call a new "PossiblyWriteLabAssignmentsToRobot"
- In CozmoExperiments, PossiblyWriteLabAssignmentsToRobot compares the
  last set of assignments read in from robot, to the current set, and only writes
  the current set to robot if it is different

* COZMO-14548 Audio choppy: Edge case for experiment writing to robot

Since we also can potentially activate (or end) an experiment when un-backgrounding
the app, we should also detect when that happens and can then write to robot.  Note:
In the prior version, we were *always* writing experiment data to robot when un-
backgrounding the app, which is wasteful.  This fixes that.
However, when an experiment becomes un-activated (due to passing the end date of
the experiment) while the user is back-grounded, we don't detect this change, because
the AnkiLab code only registers on activations, not de-activations.  But this is a non-
issue, and the robot-written experiments list would be corrected on the next robot
connection.

- Call PossiblyWriteLabAssignmentsToRobot in un-backgrounding, after trying to activate
  experiment(s)
- PossiblyWriteLabAssignmentsToRobot now makes a copy of its baseline data, so that
  when we call it in the future it detects a change