COZMO-15305 Restore button not available for 2nd robot (#6245)
Since the second robot never goes through onboarding, the onboarding flag on the robot is not set. This means we will not be saving backups for the robot. There needed to be 2 or more backup files in order to show the Restore Cozmo button and since we would never save a backup for the second robot, the user would only have one backup file for the 1st robot.

The fix to get the restore button to show for the second robot is to just always show it as long as there are one or more backup files. This means that a user with only one Cozmo will see the Restore button but it won't really do anything when pressed