COZMO-15358 Code Lab: Recreate Magic Fortune Teller ProjectJSON #6289
Workaround bug where a unicode character gets introduced at runtime, then causes the project to not load on Android. Shawn is looking into the actual fix.

Here, the change was to throw away the old operator equals block that was supposed to check if variable = 5 but was coming up empty, and regenerate that operator equals block.