COZMO-14390 Code Lab vertical: remove glow filter so project doesn't disappear (#6298)
Turn off vertical stack glow when script is executing. This is a Gaussian filter applied to an SVG, and it causes vertical programs to hide sometimes when they are running.

Fixes COZMO-14390 and COZMO-14663 (tested by QA).

Adam is still looking into a fix but hopes of finding one currently do not look good.

We will need to find an alternative in the future.

Amy and Lauren have signed off on removing it for the 2.1 release.