VIC-519 Configure WiFi over BLE via CLI
victor-ble-cli is a simple command line interface to scan for and talk
to Victors over BLE.  It is based on the work of Bryon Vandiver in
Cozmo (see robot/ble-tool) in the cozmo-one repo.  See the README.md
file for an example of how to use this tool to configure a new WiFi
access point on a Victor and have it get an IP address via DHCP.

This tool is intended for developer/tester use and NOT for end users.