COZMO-15444 Prevent NaNs being sent to JS in CozmoState (#6373)
The Json serializer in C# incorrectly serializes NaN float values as NaN (without quotes), which is invalid JSON and breaks the Javascript side.
To avoid this, we now sanitize all float values sent in the CozmoState, so that NaNs and infinity are sent as 0.0f instead. This also simplifies the Vector2 and Vector3 structs sent to Cozmo to remove redundant extra fields from the Unity versions.

This also fixes the bad log seen in COZMO-15440 , although the specifics of that bug are unrelated.