VIC-629 victor-ble-cli - fix date on Victor
Victor's clock gets reset on every reboot to Jan 1, 1970.  This change
adds a new command, `sync-time`, that will set Victor's clock to the
host time.  Also, at connection time, we will request the current year
from Victor and if it is 1970, we will assume this is incorrect and
automatically sync the time.