COZMO-15519 Code Lab: fix What's New for non-English locales (#6417)
Set all localized projects to have the same UUID as for English. This is necessary for the featured projects since the What's New modal only associates 1 UUID per project. I also made the same change for sample project Laser Smile though technically it may not yet be necessary there.

When the Lists _CodeLabSampleProjects and _CodeLabFeaturedProjects are first populated, remove projects that are specifically for other locales.