VIC-458 Remove mediandk & camera2ndk deps (#228)
After applying patches to minimize the system image size, mediandk and
camera2ndk are no longer included.  We don't have any code that we are
currently using that depends on them. This change removes them from the
build system so that the engine does not crash trying to load the
shared libs.