COZMO-15531 Code Lab: Fix German Quit button text word break (#6419)
Prior to this fix, the button pill text inherits CSS word-wrap property break-word. For text in the modal (not the button text itself), we do want to allow text to break in the middle of the word if the user creates a variable with a super long variable name with no space. But for button text, which is always predefined by Anki (typically "OK" and "Cancel"), we don't ever want to break in the middle of the word.