Mwillett/vic 614 Truncate Unsafe Plans (#226)
Allows the planner to reset undoled path segments to the robot thread if it decides some later section of the currently executed path is not safe. Additionally, if the already doled path segments are unsafe, EStop the robot.

Additionally adds a start angle to the point turn path segment, as well as GetStartAngle for all path segments, so robot start position is deterministic for paths.