VIC-629 : BLE Time Server for Victor and friends
Advertise the current time service and the "anki" time service.  Current
Time Service tries to follow the Bluetooth SIG spec to deliver time in
their funky 256bit format.

The "Anki" time service just gives you the number of milliseconds since
the epoch (Jan 1, 1970 UTC) in text format and the name of the current
time zone (like America/Los_Angeles) in text format.

The idea is to allow robots who boot up in the immediate area and can
act as a BLE central to be able to self configure the correct time
without needing WiFi to be active or to have an IP address.