VIC-694: Allow lift motor to be used while on charger (#293)
Automatically enables motor right before commanding motion and disables right after so that charging can resume afterwards. This means that the motor's position is not constantly held after the target angle is reached and instead it goes limp. The effects of the EnableMotorPower message are honored in that if the motor is disabled with the msg, subsequent lift commands are also ignored while on charger and the motor is not automatically re-enabled when removed from charger.