COZMO-15703 Restore robot needs levels after skipping onboarding (#6477)
* COZMO-15703 Restore robot needs levels after skipping onboarding

If the user skips onboarding, this makes it so the needs levels, stars, etc.
are reset to what they were when the robot was read.  If the user does not
skip onboarding (even on a second device with a freshly-installed app), we
do not restore, but rather initialize (as we did prior to this change.)

- Add bool 'onboardingSkipped' to RegisterOnboardingComplete GtE
  message, and set it appropriately on game side
- Engine side, in InitReset, if this flag is set, do the new thing which is to
  copy the needs state from the version that was read from robot upon
  connection
- Also don't award the 'first token' that's given when onboarding is completed,
  if the user skipped onboarding

* COZMO-15703 More: Fix bug: Token reward dialog coming up at wrong times

This bug has to do with the number of tokens (aka stars) the user has, and
using one robot with two devices.  Previously:  If you connected to a robot on
device #1, and have 1 star, then connected to the same robot on device #2,
and awarded an additional star (resulting in 2 stars), then connect to the
same robot on device #1 again, then when the needs hub view was displayed,
the 'you just got a star' dialog appeared (inappropriately; it should only appear
when you actually get the star.)

- Engine side:  NeedsManager::InitAfterReadFromRobotAttempt:  Send the
  existing RobotChangedFromLastSession message also when we have just
  got data from the robot that is newer than the device data
- Unity side:  NeedsStateManager:  kRobotChangedFromLastSession ->
  HasRobotChangedFromLastSession (k indicates constant and this is not a
  constant; also disambiguating from the message/handler names for easier
  reading)
- Also add missing RemoveCallback for the handler
- Unity side:  DataPersistenceManager:  Use accessor to set DisplayedStars
  so we get the profile save to occur
- Also remove redundant clearing of kRobotChangedFromLastSession bool