VIC-723 robot camera perf (#312)
* VIC-723 Use RDI stream to acquire raw bayer images

The RDI stream seems to eliminate use of the hardware ISP, which currently
is a source of kernel lock contention in our version of the OS. This change
reads raw camera data directly in a 10bpp MIPI Bayer format.  The downsampling
method in the camera system changed to accommodate this format, and there
is now a separate utility in the test folder for converting raw bayer image
dumps to tiff format.

Related minor fixes:

- Add clean_invalidate_buf method to stream config

- Fix callback type def in camera_test_main

- Silence proto_camera logs by default.

- Add CLI options to camera_test program (optionally dump images)

* VIC-723 Compile proto_camera lib with optimization

The only code in proto_camera is the camera config + downsample
processing.  Most developers will not be debugging this code and without
the hardware ISP, we need optimization to run the downsampling code
as fast as possible.

* VIC-723 Run robot_supervisor on isolated CPU

As of OS version 235 (DVT1 release candidate), we have isolated
a CPU core from the linux scheduler. Since it is no longer in the
pool, we should use this CPU to run the robot process to avoid system
load pressure from causing the scheduler to potentially delay running
the process.