Mwillett/cozmo 15756 retry roll block if planner fails (#6493)
If the goal is in an obstacle when planning, a BFS is done to try and find the closest valid pose to the obstacle. If the obstacle is too large, this can be very slow, so we limit the number of expansions. We used to report this as a failure, but instead, we now just return the total cost accumulated so far as a lower bound on the true cost of driving into a soft obstacle.