COZMO-15521 Code Lab: Prevent "Save as Remix?" dialog when user only has run program (#6506)
Don't show "Save as Remix?" prompt when user has done nothing but run the program.

When running the program, existing variable values can get changed when the variable set or change blocks are run. Afterwards, the serialized program is actually different due to those variable changes. However, we want to avoid the user being presented the "remix?" dialog in this case.