COZMO-15767 Code Lab: fix crash on import (#6513)
Prior to this change, in Cozmo version 2.2, when a Code Lab file is opened when the app is not yet launched on Android, the app crashes. This occurs due to method openCodelabFileFromURI (which is called from onNewIntent) making DAS calls. This did not happen with 2.1 for some reason that I have not tracked down. Instead of logging errors to DAS at this very early point in the launch, add the errors to the CodeLabGame list _importErrors.