VIC-787: LocalUdpServer disconnect fix in Webots (#346)
If Webots process slows down (like when attaching with debugger or computer is bogged down with other stuff) the robot process can end up trying to send enough mic data to flood the send buffer (since it's captured in real-time vs. webots time). Because of the bursty nature or audio capture we limit the amount of data that can be sent over a 6 tic window. This also allows the send buffer size to be the same for both real and sim.