COZMO-15770 Needs levels bug after reinstall/skip onboarding (#6521)
The repro of this bug:  Ensure you have 'fist bump' unlocked.  (This one way we
'detect' an old robot.)  Corrupt the robot's needs save (by quickly killing the app
immediately after backgrounding, to interrupt the save).  Then un-install the app,
and re-install the app.  When given the option, SKIP onboarding.  At that point
you get to the needs hub view with 0% needs levels, and no broken parts, so
the user is forced to erase the robot.
The cause of this bug:  In the (relatively recent) logic to handling onboarding
skipped, we were setting needs state from what was loaded from the robot.
But if that has been corrupted, it's garbage.  The fix is to fall back to fresh needs
data (the only thing we can do at this point, since the old robot's needs levels
are corrupt on the robot and the app has been un-installed so it's not on device
either.)

- NeedsManager::InitReset:  Add a check for the robot having had valid needs
  data before trying to copy from it
- Also, send needs state to game when neither robot or device has data, to
  ensure game has current state
- Add more logging