COZMO-15326 Fix explorer mode head bobbing on busy floor (#361)
The COZMO-15326 fix that we put in per https://github.com/anki/cozmo-one/pull/6467 was to:
(1) Add support to `AnimationGroup::GetAnimationName()` for a strict cooldown time. The boolean for that defaults to false, but when it is set to true, that method will NOT simply choose the animation closest to being off cooldown when all animations in the group are on cooldown.
(2) Add a cooldown time of 3.0 seconds to all the animations in the `ag_explorer_huh` animation group (so that animation is not triggered too frequently on our busy 18th floor carpet).
(3) Update the trigger for playing the cliff event animation for drone/explorer mode so it doesn't fire if that action is already running (so this animation cannot interrupt itself if multiple "cliffs" are detected in rapid succession).

We are now bringing the strict cooldown change (1 and 2) to Victor.  For Victor, we will use a different solution for 3 than we used for Cozmo. The current plan is to add support for a "canBeInterrupted" flag or something similar to actions for Victor. That is tracked in jira VIC-796