Cosmetic changes to binary_conversion.py (#6536)
The project/buildScripts/binary_conversion.py script was forked for Victor in https://github.com/anki/victor/pull/336 for VIC-103. I'm making some minor cosmetic changes in this Cozmo version to help support that fork and make it easy to compare both versions of this script (since we will be supporting both for the foreseeable future).