Add Go build system, cloud process code (#350)
* 1.) Add build support for Go to Victor

- `anki_build_go.cmake` provides utilities for building Go code;
  it sets up the proper environment for the platform, then runs
  `go get` to fetch dependencies and `go build` to generate
  binaries. Output can either be executables or static libraries
  to be linked with other C code.
- Modifies `metabuild.py` and `anki_build_source_list.cmake` to
  support new formats for Go projects; rather than specifying
  sources, Go projects build out of a folder, and create a
  dependency list (so cmake knows when to rebuild) consisting of
  everything in that folder as well as everything in the central
  Go path that stores shared dependencies. Unfortunately, this
  means that all Go code is rebuilt anytime any dependencies are
  changed; I experimented with generating lists of direct
  dependencies for each project, but that was _way_ slower (at
  configure time) than simply rebuilding Go stuff, which is pretty
  fast.
- Adds `go` to the list of dependencies in fetch-build-deps
- Sets up Android toolchain in `CMakeLists.txt` to pass to the Go
  build

* 2.) Add shared Go code

The `cloud/go` directory is the central GOPATH for building cloud
code, and will contain all packages that might be referenced by
other libraries or executables. Currently this includes...
- the `ipc` package, which provides numerous implementations for
  IPC over various protocols that should be compatible with our
  corresponding C++ implementations
- the `util` package which provides a couple small helper items

* 3.) Add cloud process code

This is the code for the cloud process, which runs on the robot
getting audio from the microphone (after the trigger word happens)
and sends it to the cloud Chipper service, which eventually returns
back an intent to send to the AI.

The process itself is in a shared area so that other test harness
programs can utilize it.

* 4.) Add cloud/voice testing programs

voicetest is an OSX command line program that records audio from
the laptop microphone and sends it to the Chipper service, for
testing voice intents. proctest takes this one step further by
providing replacement Go code for voicetest that directly utilizes
the cloud process code in `go/src/anki/cloudproc`, faking the
communication channels to and from the cloud process, in order to
directly test our on-robot stack while doing this voice testing.

* 5.) Add IPC testing/benchmarking programs

These programs were written for test purposes to ensure interop
compatibility between our C++ and Go IPC communications, and for
benchmarking the latency time of sending large chunks of data
through various IPC communication mechanisms.

* 6.) Set up cloud build

This modifies project files to add the cloud subdirectory as part
of our build, and sets up the executables to be built as part of
the cloud project, using the Go cmake tools added in earlier
commits.

* 7.) Add custom property for target output

Since Go executables are custom targets, passing them to the
`android_strip` function will break because the `$<TARGET_FILE>`
generator expression only works with executable and library
targets. This adds a custom property, `ANKI_OUT_PATH`, that can be
set on targets to manually specify their output location, and then
be checked in `android_strip` for use as an alternative to
`TARGET_FILE`. The property will be automatically set on Go
executable targets.

* Build config updates

- make chipper key part of DEPS
- add support for curling individual files to dependencies.py
- add local config to always use SSH for github instead of HTTPS
- move responsibility for chipper key into cloud's CMakeLists.txt
  instead of indiscriminately applying it to every Go target

* Local repo option doesn't seem to work, try global