COZMO-15916 Code Lab vertical: fix project not loading due to dupe vars #6582
Fixes COZMO-15916 and COZMO-15917 (which likely are dupe tickets).

Instead of throwing an error when a variable name is encountered more than one time, now we continue to load the project. This fixes the issue where the project does not load at all but may still run when you tap the green flag button.

This issue comes up with variables named "undefined". QA has reproduced this bug when renaming and deleting variables.

Unfortunately, even this fix, if the user tries to delete the "undefined" variable, it still comes back when the project is next loaded, as Mark points out.

I am hoping that when we update to the new scratch-blocks and scratch-vm libraries in Cozmo app 2.3 that this issue will be fixed but I have not yet confirmed this is the case.

With this fix, all known files that encountered this bug load, including the following:

* ETC colorswitch (colorswitch.codelab)
* forum user issue (remix_of_my_project_7.codelab)
* QA reported file (my_project_52.codelab)