VIC-901: Lift and head actions sometimes fail with MOTOR_STOPPED_MAKI… (#417)
* VIC-901: Lift and head actions sometimes fail with MOTOR_STOPPED_MAKING_PROGRESS even though they end up working just fine.

This is because the action sometimes prematurely thinks that the motor has started executing its motor command when in fact it's still moving from a previous action's motor command. During the transition to the current action's motor command, it thinks it has stopped moving and since it's not yet in position it reports this failure, after which the motor moves to the commanded position just fine.

This fix makes sure that the motor has stopped moving before the motor command is sent so the CheckIfDone() doesn't prematurely fail.