SearchForNearbyObjectAction failing with MOTOR_STOPPED_MAKING_PROGRESS (#458)
SearchForNearbyObjectAction had track locking suppressed on its compound action so no tracks (beyond the BODY track which is locked in the ctor) was actually locked. This meant that animations played in parallel with SearchForNearbyObjectAction could interrupt the MoveToHead sub-action causing it to fail with MOTOR_STOPPED_MAKING_PROGRESS.

Removed track lock suppression so now sub-actions are responsible for their own track locking.