VIC-1035 Set date/time on Victor running LA or LE
On Victors running LE, we need to pass different arguments to the date
command to get it set.  To preserve compatibility with both LA and LE,
first send the old args for LA and then immediately send the new args
for LE.  For LE the first command will fail, but the second will
succeed.  For LA, the reverse is expected.