Add verbose logging to cloud process (#498)
* Add verbose logging to cloud process

If started with the "-verbose" flag, the cloud process will add
more logging about when it receives data from the mic, when that
data is uploaded, and when an intent is received from the server.
It does not (yet) add more logging to Chipper, which would probably
be needed to get a complete picture of networking costs.

* Improve voice capture testing program

- Fix a bug that caused the written wave file to not have any
  samples in it (audio callback was tossing samples after upload)
- Put upload on a separate routine behind a buffered channel to
  prevent audio recording from getting blocked while waiting for
  upload to happen
- Add voice sample to be used in future

* Add mechanism to test cloud client on-robot w/o mics

- Add wavtest program, which will read in a wav file (specified on
  command line) and feed the samples inside it to the cloud client
  through a test socket as if the data was being recorded in real
  time
- Add test server socket to cloud client to accept connections from
  wav tester

* Temporarily disable wavtest build target

Having multiple go targets built by default causes issues with
`go get` running in parallel and fetching the same remote
repositories to put into the same location. Until a better fix for
this is added, we don't need to build wavtest by default.